Added the Feature to Share Vidoes in the Discussion Forum
