+++ lib/screen/discussion/emoji_picker.dart
@@ -1,0 +1,382 @@
+import 'package:flutter/material.dart';
+import '../../theme/theme_provider.dart';
+
+class EmojiPickerWidget extends StatelessWidget {
+  final ThemeProvider themeProvider;
+  final AnimationController emojiAnimationController;
+  final Animation<double> emojiScaleAnimation;
+  final String selectedEmojiCategory;
+  final Function(String) onCategorySelected;
+  final Function(String) onEmojiSelected;
+
+  // Emoji categories and data
+  static const Map<String, List<String>> _emojiCategories = {
+    'Smileys': [
+      '😀',
+      '😃',
+      '😄',
+      '😁',
+      '😆',
+      '😅',
+      '😂',
+      '🤣',
+      '😊',
+      '😇',
+      '🙂',
+      '🙃',
+      '😉',
+      '😌',
+      '😍',
+      '🥰',
+      '😘',
+      '😗',
+      '😙',
+      '😚',
+      '😋',
+      '😛',
+      '😝',
+      '😜',
+      '🤪',
+      '🤨',
+      '🧐',
+      '🤓',
+      '😎',
+      '🤩',
+      '🥳',
+      '😏',
+      '😒',
+      '😞',
+      '😔',
+      '😟',
+      '😕',
+      '🙁',
+      '☹️',
+      '😣',
+      '😖',
+      '😫',
+      '😩',
+      '🥺',
+      '😢',
+      '😭',
+      '😤',
+      '😠',
+      '😡',
+      '🤬',
+    ],
+    'Hearts': [
+      '❤️',
+      '🧡',
+      '💛',
+      '💚',
+      '💙',
+      '💜',
+      '🖤',
+      '🤍',
+      '🤎',
+      '💔',
+      '❣️',
+      '💕',
+      '💞',
+      '💓',
+      '💗',
+      '💖',
+      '💘',
+      '💝',
+      '💟',
+    ],
+    'Gestures': [
+      '👍',
+      '👎',
+      '👌',
+      '🤌',
+      '🤏',
+      '✌️',
+      '🤞',
+      '🤟',
+      '🤘',
+      '🤙',
+      '👈',
+      '👉',
+      '👆',
+      '🖕',
+      '👇',
+      '☝️',
+      '👋',
+      '🤚',
+      '🖐',
+      '✋',
+      '🖖',
+      '👏',
+      '🙌',
+      '🤲',
+      '🤝',
+      '🙏',
+    ],
+    'Objects': [
+      '🎉',
+      '🎊',
+      '🎈',
+      '🎂',
+      '🎁',
+      '🎀',
+      '🏆',
+      '🏅',
+      '🥇',
+      '🥈',
+      '🥉',
+      '⚽',
+      '🏀',
+      '🏈',
+      '⚾',
+      '🥎',
+      '🎾',
+      '🏐',
+      '🏉',
+      '🥏',
+      '🎱',
+      '🪀',
+      '🏓',
+      '🏸',
+      '🏒',
+      '🏑',
+      '🥍',
+      '🏏',
+      '🪃',
+      '🥅',
+      '⛳',
+      '🪁',
+      '🏹',
+      '🎣',
+      '🤿',
+      '🥊',
+      '🥋',
+      '🎽',
+    ],
+    'Nature': [
+      '🌞',
+      '🌝',
+      '🌛',
+      '🌜',
+      '🌚',
+      '🌕',
+      '🌖',
+      '🌗',
+      '🌘',
+      '🌑',
+      '🌒',
+      '🌓',
+      '🌔',
+      '🌙',
+      '🌎',
+      '🌍',
+      '🌏',
+      '🪐',
+      '💫',
+      '⭐',
+      '🌟',
+      '✨',
+      '⚡',
+      '☄️',
+      '💥',
+      '🔥',
+      '🌪',
+      '🌈',
+      '☀️',
+      '🌤',
+      '⛅',
+      '🌦',
+      '🌧',
+      '⛈',
+      '🌩',
+      '🌨',
+      '❄️',
+      '☃️',
+      '⛄',
+      '🌬',
+    ],
+    'Food': [
+      '🍎',
+      '🍐',
+      '🍊',
+      '🍋',
+      '🍌',
+      '🍉',
+      '🍇',
+      '🍓',
+      '🫐',
+      '🍈',
+      '🍒',
+      '🍑',
+      '🥭',
+      '🍍',
+      '🥥',
+      '🥝',
+      '🍅',
+      '🍆',
+      '🥑',
+      '🥦',
+      '🥬',
+      '🥒',
+      '🌶',
+      '🫑',
+      '🌽',
+      '🥕',
+      '🫒',
+      '🧄',
+      '🧅',
+      '🥔',
+      '🍠',
+      '🥐',
+      '🥖',
+      '🍞',
+      '🥨',
+      '🥯',
+      '🧀',
+      '🥚',
+      '🍳',
+      '🧈',
+      '🥞',
+      '🧇',
+      '🥓',
+      '🥩',
+      '🍗',
+      '🍖',
+      '🦴',
+      '🌭',
+      '🍔',
+      '🍟',
+    ],
+  };
+
+  const EmojiPickerWidget({
+    super.key,
+    required this.themeProvider,
+    required this.emojiAnimationController,
+    required this.emojiScaleAnimation,
+    required this.selectedEmojiCategory,
+    required this.onCategorySelected,
+    required this.onEmojiSelected,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimatedBuilder(
+      animation: emojiAnimationController,
+      builder: (context, child) {
+        return Transform.scale(
+          scale: emojiScaleAnimation.value,
+          alignment: Alignment.bottomCenter,
+          child: Container(
+            height: MediaQuery.of(context).viewInsets.bottom > 0
+                ? MediaQuery.of(context).viewInsets.bottom
+                : 280,
+            decoration: BoxDecoration(
+              color: themeProvider.isDarkMode ? Colors.grey[800] : Colors.white,
+              borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.black.withOpacity(0.1),
+                  blurRadius: 10,
+                  offset: Offset(0, -2),
+                ),
+              ],
+            ),
+            child: Column(
+              children: [
+                // Drag indicator
+                Container(
+                  margin: EdgeInsets.only(top: 8),
+                  width: 40,
+                  height: 4,
+                  decoration: BoxDecoration(
+                    color: Colors.grey[400],
+                    borderRadius: BorderRadius.circular(2),
+                  ),
+                ),
+
+                // Category tabs
+                Container(
+                  height: 50,
+                  child: ListView(
+                    scrollDirection: Axis.horizontal,
+                    padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                    children: _emojiCategories.keys.map((category) {
+                      final isSelected = selectedEmojiCategory == category;
+                      return Padding(
+                        padding: EdgeInsets.only(right: 8),
+                        child: GestureDetector(
+                          onTap: () => onCategorySelected(category),
+                          child: Container(
+                            padding: EdgeInsets.symmetric(
+                                horizontal: 16, vertical: 6),
+                            decoration: BoxDecoration(
+                              color: isSelected
+                                  ? Color(0xFF2196F3)
+                                  : (themeProvider.isDarkMode
+                                      ? Colors.grey[700]
+                                      : Colors.grey[200]),
+                              borderRadius: BorderRadius.circular(20),
+                            ),
+                            child: Text(
+                              category,
+                              style: TextStyle(
+                                color: isSelected
+                                    ? Colors.white
+                                    : (themeProvider.isDarkMode
+                                        ? Colors.grey[300]
+                                        : Colors.grey[600]),
+                                fontWeight: isSelected
+                                    ? FontWeight.bold
+                                    : FontWeight.normal,
+                                fontSize: 12,
+                              ),
+                            ),
+                          ),
+                        ),
+                      );
+                    }).toList(),
+                  ),
+                ),
+
+                // Emoji grid
+                Expanded(
+                  child: GridView.builder(
+                    padding: EdgeInsets.all(16),
+                    gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+                      crossAxisCount: 8,
+                      childAspectRatio: 1,
+                      crossAxisSpacing: 8,
+                      mainAxisSpacing: 8,
+                    ),
+                    itemCount: _emojiCategories[selectedEmojiCategory]!.length,
+                    itemBuilder: (context, index) {
+                      final emoji =
+                          _emojiCategories[selectedEmojiCategory]![index];
+                      return GestureDetector(
+                        onTap: () => onEmojiSelected(emoji),
+                        child: Container(
+                          decoration: BoxDecoration(
+                            color: themeProvider.isDarkMode
+                                ? Colors.grey[700]
+                                : Colors.grey[100],
+                            borderRadius: BorderRadius.circular(8),
+                          ),
+                          child: Center(
+                            child: Text(
+                              emoji,
+                              style: TextStyle(fontSize: 20),
+                            ),
+                          ),
+                        ),
+                      );
+                    },
+                  ),
+                ),
+              ],
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
